move getDefaultValues to api  and update imports
